add new cmdlet and move methods from sbManager
